Fix actions tests (#5)
* Fix actions tests

* More Tests
