+++ package.json
@@ -62,7 +62,7 @@
     "@fortawesome/react-fontawesome": "^0.1.7",
     "@primer/octicons-react": "^9.3.1",
     "auto-launch": "=5.0.1",
-    "axios": "=0.16.1",
+    "axios": "=0.19.0",
     "bootstrap": "=4.0.0-alpha.6",
     "css-loader": "^3.2.1",
     "date-fns": "^2.8.1",
@@ -104,12 +104,11 @@
     "@types/react": "^16.9.16",
     "@types/react-redux": "^7.1.5",
     "@types/styled-components": "^4.4.1",
-    "babel-jest": "=20.0.3",
     "electron": "=7.1.3",
     "electron-packager": "=8.7.0",
     "enzyme": "=2.8.2",
     "jest": "^24.9.0",
-    "nock": "=9.0.13",
+    "nock": "=11.7.0",
     "prettier": "=1.19.1",
     "react-test-renderer": "=16.12.0",
     "redux-logger": "=3.0.6",
