--- src/js/__tests__/components/logos/white.js
@@ -1,24 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import { shallow } from 'enzyme';
-
-import LogoWhite from '../../../components/logos/white';
-
-describe('components/logos/white.js', () => {
-  it('renders correctly', () => {
-    const tree = renderer.create(<LogoWhite />);
-
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('should click on the logo', () => {
-    const onClick = jest.fn();
-    const wrapper = shallow(<LogoWhite onClick={onClick} />);
-
-    expect(wrapper).toBeDefined();
-
-    wrapper.find('svg').simulate('click');
-
-    expect(onClick).toHaveBeenCalledTimes(1);
-  });
-});
