+++ src/js/components/all-read.test.tsx
@@ -4,6 +4,8 @@ import * as TestRenderer from 'react-test-renderer';
 import Constants from '../utils/constants';
 import { AllRead } from './all-read';
 
+jest.mock('react-typist');
+
 describe('components/all-read.tsx', function() {
   it('should render itself & its children', function() {
     spyOn(Constants, 'ALLREAD_EMOJIS');
