+++ src/js/components/logos/white.tsx
@@ -13,6 +13,7 @@ export const LogoWhite = (props: IProps) => {
       xmlns="http://www.w3.org/2000/svg"
       xmlnsXlink="http://www.w3.org/1999/xlink"
       viewBox="0 0 180.31 182.52"
+      role="logo"
     >
       <defs>
         <linearGradient
